+++ app/components/productsFilterList/productsFilterList.js
@@ -1,0 +1,10 @@
+/**
+ * Created by pua on 9/15/2016.
+ */
+angular.module("iAngelsTeaShop")
+    .directive("productsFilterListDirective", function() {
+    return {
+        templateUrl : "app/components/productsFilterList/productsFilterList.html"
+    };
+});
+